OrderComponent/requirements.txt
googlemaps
Flask==2.0.2
Flask_RESTful==0.3.9
Flask_Bcrypt==0.7.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
Werkzeug==2.0.0
PyJWT==2.4.0
python-decouple